Add initial setup for backend with database connection and swagger
This commit includes initial configurations for the backend of the application. It introduces sequelize for handling database communication and adds a setup for connecting to a SQLite database with some dummy data. In addition, a basic implementation of user routes and controllers has been added. Swagger for API documentation is set up alongside JSDoc for additional documentation. Necessary dependencies have been added to the package.json file.