Update ESLint rules and make various UI improvements
Disables the 'no-explicit-any' rule in ESLint configuration. Adds new functionality to the Hero component with the addition of a button toolbar, adds new pages for each CV based on the user's language. Updates styling in App.css including button and card colors, and responsiveness. Also, introduces a new Skills component.